Refactor codecs package
