+++ mlserver/codecs/__init__.py
@@ -1,0 +1,22 @@
+from .numpy import NumpyCodec, NumpyRequestCodec
+from .string import StringCodec
+from .base import (
+    InputCodec,
+    RequestCodec,
+    register_input_codec,
+    register_request_codec,
+    find_input_codec,
+    find_request_codec,
+)
+
+__all__ = [
+    "NumpyRequestCodec",
+    "NumpyCodec",
+    "StringCodec",
+    "InputCodec",
+    "RequestCodec",
+    "register_input_codec",
+    "register_request_codec",
+    "find_input_codec",
+    "find_request_codec",
+]
