+++ tests/codecs/test_numpy.py
@@ -1,28 +1,8 @@
 import pytest
 import numpy as np
 
-from mlserver.codecs import StringCodec, NumpyCodec
+from mlserver.codecs import NumpyCodec
 from mlserver.types import RequestInput
-
-
-@pytest.mark.parametrize(
-    "encoded, payload",
-    [
-        (b"hello world", "hello world"),
-    ],
-)
-def test_string_codec(encoded, payload):
-    request_input = RequestInput(name="foo", shape=[], datatype="BYTES", data=encoded)
-    codec = StringCodec()
-    decoded = codec.decode(request_input)
-
-    assert decoded == payload
-
-    response_output = codec.encode(name="foo", payload=decoded)
-
-    assert response_output.datatype == "BYTES"
-    assert response_output.shape == [len(encoded)]
-    assert response_output.data.__root__ == encoded
 
 
 @pytest.mark.parametrize(
